Refactor to have single definition of spin rotation.
