+++ lib/measurements/baryon_contractions_w.cc
@@ -15,6 +15,7 @@
 #include "../momentum/lalibe_sftmom.h"
 #include "meas/inline/io/named_objmap.h"
 #include "io/qprop_io.h"
+#include "util/spin_basis.h"
 
 #include <set>
