+++ lib/measurements/meson_contractions_w.cc
@@ -16,6 +16,7 @@
 #include "../momentum/lalibe_sftmom.h"
 #include "meas/inline/io/named_objmap.h"
 #include "io/qprop_io.h"
+#include "util/spin_basis.h"
 
 
 namespace Chroma
